Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-scrolled-viewport-expected.txt


FAIL Tests modal dialog's containing block is the initial containing block assert_equals: expected null but got Element node <div style="position: absolute; top: 5000px; left: 5000px...
